+++ src/config.js
@@ -20,7 +20,6 @@ const fileTypes = {
     "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
     "application/vnd.openxmlformats-officedocument.presentationml.presentation",
     "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-    "application/pdf",
   ],
   oldOffice: ["application/msword", "application/vnd.ms-excel", "application/vnd.ms-powerpoint"],
   officeConvert: {